requirements.txt
certifi==2017.11.5
chardet==3.0.4
coverage==4.4.2
funcsigs==1.0.2
idna==2.6
nose==1.3.7
pbr==3.1.1
pycurl==7.43.0
python-dateutil==2.6.1
requests==2.18.4
six==1.11.0
urllib3==1.22
